tcc_cc.c: Some clean-up and bug fix
tcc_cc.c implemented some functions from the standard library. Replaced
those the standard library ones.

Fixed a bug with respect to generating constant values in the Stack-C output.
These are now printed as unsigned numbers, because there is problem with
printing the smallest 32 bit signed value with a function that is limited to
32 bit values.

In the make file tcc_s is compiled with the tcc_cc_s (instead of tcc_cc),
which is tcc_cc.c compiled with tcc_cc (compiled with GNU C).

Some small edits in tcc_cc.md and adding to test to unittest.c
that should have been commit with the previous commit.